requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers
python-dotenv
